1. wizard function params. 2. check params before and after request soap service

+++ protected/models/SoapFunction.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * Модель для таблицы soap_function.
  * Список функций, которые поддерживает определенный SOAP сервис.
@@ -16,6 +15,11 @@
  */
 class SoapFunction extends CActiveRecord
 {
+    const FUNCTION_TYPE_GET = 'get';
+    const FUNCTION_TYPE_LIST = 'list';
+    const FUNCTION_TYPE_SAVE = 'save';
+    const FUNCTION_TYPE_DELETE = 'delete';
+
 	/**
 	 * Returns the static model of the specified AR class.
 	 * @param string $className active record class name.
@@ -32,6 +36,25 @@ class SoapFunction extends CActiveRecord
     public function tableName()
     {
         return 'soap_function';
+    }
+
+    /**
+     * @param integer $input_param
+     * @return SoapFunctionParam[] Возвращает список входных параметров. Формат [name => model].
+     * @throws CHttpException
+     */
+    public function getParamsByType($input_param = SoapFunctionParam::TYPE_INPUT)
+    {
+        /**
+         * @var $params SoapFunctionParam[]
+         */
+        $params = array();
+        foreach($this->soapFunctionParams as $p){
+            if ($p->input_param == $input_param){
+                $params[$p->name] = $p;
+            }
+        }
+        return $params;
     }
 
     /**
@@ -84,26 +107,183 @@ class SoapFunction extends CActiveRecord
     /**
      * @param mixed $val
      * @param string $type
-     * @param bool $is_empty
-     * @return bool
+     * @param boolean $required
+     * @return boolean
      */
-    private function _checkType($val, $type, $is_empty=true)
+    private function _checkType($val, $type, $required = false)
     {
-        if ($is_empty && empty($val)){
-            return true;
+        if ($required){
+            if ($type == SoapFunctionParam::TYPE_DATA_BOOLEAN){
+                if (!is_bool($val) && empty($val)){
+                    return false;
+                }
+            } elseif (empty($val)){
+                return false;
+            }
         }
-        if ($type == 'string'){
+
+        if ($type == SoapFunctionParam::DEFAULT_TYPE_OF_DATA){
             return is_string($val);
-        } elseif ($type == 'integer'){
+        } elseif ($type == SoapFunctionParam::TYPE_DATA_INTEGER){
             return is_int($val);
-        } elseif ($type == 'bool'){
+        } elseif ($type == SoapFunctionParam::TYPE_DATA_BOOLEAN){
             return is_bool($val);
-        } elseif ($type == 'array'){
+        } elseif ($type == SoapFunctionParam::TYPE_DATA_ARRAY){
             return is_array($val);
-        }elseif ($type == 'date'){
+        }elseif ($type == SoapFunctionParam::TYPE_DATA_DATE){
             return (FALSE !== strtotime($val));
         }
         return false;
+    }
+
+    /**
+     * @param string $return JSON строка с ответом от SOAP сервиса.
+     * @return array
+     * @throws CSoapTestException
+     */
+    public function checkAfterRequest($return)
+    {
+        $not_found = array();
+        $required = array();
+        $wrong_data_type = array();
+        if($this->type == self::FUNCTION_TYPE_DELETE){
+            if (!is_bool($return)){
+                $wrong_data_type[] = array(
+                    'key' => 'return',
+                    'type_of_data' => SoapFunctionParam::TYPE_DATA_BOOLEAN
+                );
+            }
+        } elseif($this->type == self::FUNCTION_TYPE_GET){
+            $ret = CJSON::decode($return);
+
+            if (!empty($ret)){
+                if (!isset($ret[0])){
+                    throw new CSoapTestException('Получен неизвестный результат функции.');
+                }
+                $ret = $ret[0];
+                $output_params = $this->getParamsByType(SoapFunctionParam::TYPE_OUTPUT);
+
+                foreach ($output_params as $key=>$value){
+                    if (!isset($ret[$key])){
+                        $not_found[] = $key;
+                    } elseif (!$this->_checkType($ret[$key], $output_params[$key]->type_of_data, $output_params[$key]->required)) {
+                        $wrong_data_type[] = array(
+                            'key' => $key,
+                            'type_of_data' => $output_params[$key]->type_of_data
+                        );
+                    }
+                }
+            }
+
+//            var_dump($ret);die;
+        }
+
+//        if (in_array($this->soapFunction->type, array('get', 'list'))){
+//            $ret = CJSON::decode($return);
+//            $errors = array();
+//            if ($this->soapFunction->type == 'list'){
+//                if (empty($ret)){
+//                    throw new CSoapTestException('Получен неизвестный результат функции.');
+//                }
+////                        foreach($ret as $p){
+////                            var_dump($p);die;
+//                $errors = $this->soapFunction->checkParams($ret[0]);
+////                        }
+//            } elseif ($this->soapFunction->type == 'get'){
+//                if (!empty($ret)){
+//                    if (!isset($ret[0])){
+//                        throw new CSoapTestException('Получен неизвестный результат функции.');
+//                    }
+//                    $errors = $this->soapFunction->checkParams($ret[0]);
+//                }
+//            }
+//            if (!empty($errors)){
+//                throw new CSoapTestException('Ошибки:<br/>' . implode('<br/>', $errors).'<br/>Результат:<br/>'.$return);
+//            }
+//        } elseif ($this->soapFunction->type == 'save'){
+////                    if (!ctype_digit($return)){
+////                        throw new CSoapTestException('Ошибки при сохранении данных.<br/>Результат:<br/>'.$return);
+////                    }
+////                    $args = CJSON::decode($this->args);
+////                    $ret = $this->soapFunction->checkParams($args);
+////                    if (!empty($ret)){
+////                        throw new CSoapTestException('Ошибки в передаваемых параметрах:<br/>' . implode('<br/>', $ret));
+////                    }
+//
+//
+//        } elseif($this->soapFunction->type == 'delete'){
+//            if (!is_bool($return)){
+//                throw new CSoapTestException('Ответ не boolean:<br/>Результат:<br/>'.$return);
+//            }
+//        }
+
+        if (empty($not_found) && empty($required) && empty($wrong_data_type)){
+            return array();
+        }
+        return array(
+            'not_found' => $not_found,
+            'required' => $required,
+            'wrong_data_type' => $wrong_data_type,
+        );
+    }
+
+    /**
+     * @param SoapTest $test
+     * @return array
+     * @throws CSoapTestException
+     */
+    public function checkBeforeRequest(SoapTest $test)
+    {
+        if (empty($test->args) && !is_string($test->args)){
+            throw new CSoapTestException('Не заданы аргументы теста.');
+        }
+        $args = CJSON::decode($test->args);
+        if (empty($args) || !is_array($args) || !isset($args[0])){
+            throw new CSoapTestException('Неправильный формат аргументов теста.');
+        }
+        $args = $args[0];
+
+        $input_params = $this->getParamsByType(SoapFunctionParam::TYPE_INPUT);
+        if (empty($input_params)){
+            throw new CSoapTestException('Для функции не заданы входящие параметры.');
+        }
+
+        if ($this->type != self::FUNCTION_TYPE_DELETE){
+            $output_params = $this->getParamsByType(SoapFunctionParam::TYPE_OUTPUT);
+            if (empty($output_params)){
+                throw new CSoapTestException('Для функции не заданы выходные параметры.');
+            }
+        }
+
+        $not_found = array();
+        $required = array();
+        $wrong_data_type = array();
+        if($this->type == self::FUNCTION_TYPE_SAVE){
+            // TODO для save отдельный обработчик
+        } else {
+            foreach ($args as $key=>$value){
+                if (!isset($input_params[$key])){
+                    $not_found[] = $key;
+                } elseif (empty($key)) {
+                    if ($input_params[$key]->required){
+                        $required[] = $key;
+                    }
+                } elseif (!$this->_checkType($value, $input_params[$key]->type_of_data)) {
+                    $wrong_data_type[] = array(
+                        'key' => $key,
+                        'type_of_data' => $input_params[$key]->type_of_data
+                    );
+                }
+            }
+        }
+        if (empty($not_found) && empty($required) && empty($wrong_data_type)){
+            return array();
+        }
+        return array(
+            'not_found' => $not_found,
+            'required' => $required,
+            'wrong_data_type' => $wrong_data_type,
+        );
     }
 
     /**
@@ -154,7 +334,8 @@ class SoapFunction extends CActiveRecord
                         MAX(`test_result`) AS `test_result`
                 FROM '.SoapTest::model()->tableName().'
                 GROUP BY `function_id`) t ON f.id = t.fid
-            WHERE gf.service_id = :service_id'
+            WHERE gf.service_id = :service_id
+            ORDER BY f.group_id ASC, f.name ASC'
         );
         return $cmd->queryAll(true, array(
             ":service_id" => $service_id
@@ -239,7 +420,8 @@ class SoapFunction extends CActiveRecord
 			array('type', 'in', 'range' => array_keys(SoapFunction::getTypes())),
 
             array('group_id', 'required'),
-//            array('group_id', 'in', 'range' => array_keys(SoapFunction::getTypes())),
+
+            array('SoapFunctionParam', 'validateParams'),
 
             array('description', 'length', 'max' => 45),
 
@@ -249,6 +431,42 @@ class SoapFunction extends CActiveRecord
 //			array('id, service_id, name', 'safe', 'on'=>'search'),
 		);
 	}
+
+    /**
+     * Валидация заданных параметров функции.
+     * @param string $attribute
+     */
+    public function validateParams($attribute){
+        if (isset($_POST[$attribute]) && !empty($_POST[$attribute])){
+            $input_names = array();
+            $output_names = array();
+            foreach ($_POST[$attribute] as $i=>$param){
+                $p = new SoapFunctionParam();
+                $p->attributes = $_POST[$attribute][$i];
+//                var_dump($p->attributes);die;
+                $valid = true;
+                foreach($p->attributes as $atr){
+                    if (!$p->validate($atr)){
+                        $valid = false;
+                        $this->addError($attribute.'['.$i.']['.$atr.']', $p->getError($atr));
+                    }
+                }
+                if ($p->input_param){
+                    if (!isset($input_names[$p->name])){
+                        $input_names[$p->name] = $p;
+                    } else {
+                        $this->addError($attribute.'['.$i.'][name]', 'Название входного параметра {'.$p->name.'} уже существует.');
+                    }
+                } else {
+                    if (!isset($output_names[$p->name])){
+                        $output_names[$p->name] = $p;
+                    } else {
+                        $this->addError($attribute.'['.$i.'][name]', 'Название выходного параметра {'.$p->name.'} уже существует.');
+                    }
+                }
+            }
+        }
+    }
 
 	/**
 	 * @return array relational rules.
@@ -280,22 +498,47 @@ class SoapFunction extends CActiveRecord
 	}
 
     /**
-     *  До удаление функции, удаляем все связанные с ней тесты.
+     * До удаление функции, удаляем все связанные с ней тесты {@see SoapTest}
+     * и параметры {@see SoapFunctionParam}.
      */
 	protected function beforeDelete()
 	{
-        $this->deleteTests();
-		return parent::beforeDelete();
+        if (parent::beforeDelete()){
+            $this->deleteParams();
+            $this->deleteTests();
+            return true;
+        }
+		return false;
 	}
 
     /**
-     *  Удалениям все тесты, связанные с данной функцией.
+     * Удалениям все тесты, связанные с данной функцией.
+     * @return boolean $valid
      */
     public function deleteTests()
     {
+        $valid = true;
         foreach ($this->soapTests as $t) {
-            $t->delete();
+            if (!$t->delete()){
+                $valid = false;
+            }
         }
+        return $valid;
+    }
+
+    /**
+     * Удалениям все параметры, связанные с данной функцией.
+     * @return boolean $valid
+     */
+    public function deleteParams()
+    {
+        $valid = true;
+        foreach ($this->soapFunctionParams as $t) {
+            if (!$t->delete()){
+                $valid = false;
+            }
+        }
+        return $valid;
     }
 
     /**
